refactor: each generator has it's own module
