+++ package.json
@@ -5,7 +5,8 @@
 	"scripts": {
 		"dev": "vite",
 		"build": "tsc && vite build",
-		"preview": "vite preview"
+		"preview": "vite preview",
+    "test": "jest src/__tests__/ --coverage"
 	},
 	"devDependencies": {
 		"@types/dom-to-image": "^2.6.4",
